Fixup frontend merge conflicts
